+++ kayak/__init__.py
@@ -9,3 +9,5 @@ from .kayak import GeneticEncoding
 from .kayak import GeneCode
 from .feature_types import FeatureType
 from .population import Population
+from .population import FitnessMap
+from .population import DelayedRandomFitnessMap