+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - added a population class which allows operations over sets of gene codes (concrete instances, given a gene code space description)
 - added VERSION and CHANGELOG.md files
 - fixed test errors on fit() invocations
+- added a fitness map definition and a dev-test random fitness map function with delayed timer